Add cname setting to deploy.
This makes it so the CNAME file that is committed (https://github.com/getdweb/dweb-gatsby/commit/491466ab3d49264383107f132f22d8052a9d9e75)
when setting custom domain is done automatically. We were running into broken deploys because this file is not created during a gatsby build, so it disappears everytime there is a new deploy.